Deprecate v1 create_record endpoint and fix Checks (#7)
Summary

Upgrade the Amiqus create_record endpoint to v2 and ensure that Checks are saved from their associated Records. Do this by:

Altering the data sent as part of the create_record request. See the migration guide for this endpoint for more details.

When handling the response from Amiqus at create_record, parse the result differently:

checks is no longer a key in the response data, so Checks were never being saved. Update this block to handle the v2 response

Removing support for additional kwargs to be sent in the data. This was there to add forms to the record, but these should now be added as Steps.

Why is this change being made?

So that this package needn't worry about the v1 Amiqus API, and Checks are an integral feature so they should be properly supported. It seems that you can use the Amiqus API using v1 data structures in your request, but the response uses the v2 structure, so that needs to be reflected in this endpoint.

Steps to test

Get an Amiqus Personal Access Token appropriate for local development.
Create an Amiqus Client and call the create_record helper for your Client with a list of steps according the documentation
You should now have a Record and its associated Steps

This is a breaking change and for your application to work, you must update the list of checks (steps) sent to the create_record helper