Add Step, Form, and Review models (#8)
Add support for the Step, Review and Form models from Amiqus. A set of Steps are created on Amiqus whenever a new Record is created. These can be of type Check or Form. Amiqus does also support Document type Steps but these are not needed in our use-case (PRs welcome).

Each Step can have one or more Reviews, each with their own status.

The updates in this PR are being done as to more closely match what happens on Amiqus when a Record is created.