Update parsers (#12)
Update the parsing methods on BaseModel and Check. This is because these methods were not writing properly to the database when parsing the initial data.

Make some updates to the tests as well:

Rename test_record to test_check
Rename test_report to test_record
These old names seem to be a hangover from the Onfido times.