Add helper to update to client status (#13)
Add a new helper update_client_status and a new API function patch to allow changing a Client's status in Amiqus.

This is needed because the v2 API removed the reset_client_status preference from the create_record endpoint. This would automatically reset a existing client's status to "Requires review" when a new request is created for them.

We now need to do this ourselves upon record re-creation, and this can be done using these functions.